+++ ExpandWorld/Spawn.cs
@@ -81,6 +81,8 @@ public class Spawn
     }
     var data = dataOverride(obj.Data, source, obj.Prefab);
     SetData(prefab, pos, rot, data);
+
+    //ExpandWorld.Log.LogDebug($"Spawning {obj.Prefab} at {Helper.Print(pos)} {source}");
     var go = UnityEngine.Object.Instantiate(prefab, pos, rot);
     Data.CleanGhostInit(go);
     if (ZNetView.m_ghostInit)
@@ -158,7 +160,7 @@ public class Spawn
     if (swaps.Count == 1)
       return swaps[0].Item2;
     var rng = UnityEngine.Random.value;
-    ExpandWorld.Log.LogDebug($"RandomizeSwap: Roll {Helper.Print(rng)} for {string.Join(", ", swaps.Select(s => s.Item2 + ":" + Helper.Print(s.Item1)))}");
+    //ExpandWorld.Log.LogDebug($"RandomizeSwap: Roll {Helper.Print(rng)} for {string.Join(", ", swaps.Select(s => s.Item2 + ":" + Helper.Print(s.Item1)))}");
     foreach (var swap in swaps)
     {
       rng -= swap.Item1;
@@ -173,7 +175,7 @@ public class Spawn
     if (swaps.Count == 1)
       return swaps[0].Item2;
     var rng = UnityEngine.Random.value;
-    ExpandWorld.Log.LogDebug($"RandomizeData: Roll {Helper.Print(rng)} for weigths {string.Join(", ", swaps.Select(s => Helper.Print(s.Item1)))}");
+    //ExpandWorld.Log.LogDebug($"RandomizeData: Roll {Helper.Print(rng)} for weigths {string.Join(", ", swaps.Select(s => Helper.Print(s.Item1)))}");
     foreach (var swap in swaps)
     {
       rng -= swap.Item1;